requirements.txt
djangorestframework==3.9.2
django==2.0.2
markdown==3.1
django-filter==2.1.0
pillow==6.0.0
pymysql==0.9.3
xadmin2==2.0.1